+++ app/src/main/java/garethpaul/com/androidspeaker/MainActivity.java
@@ -145,6 +145,12 @@ public class MainActivity extends Activity {
     }
 
     @Override
+    protected void onPause() {
+        releasePlayer();
+        super.onPause();
+    }
+
+    @Override
     protected void onDestroy() {
         releasePlayer();
         super.onDestroy();